requirements.txt
python-dotenv
deepgram-sdk==3.*